it is a bit of extra work, having to rely only on the standard libraries, in particular default java Date/Time/Timestamp support is really limited...
anyhow: I altered the definition of the columnTypes property: it now defines the type of the columns of the table.  the result is -by its own name- a result of the operations on the data.  calculating the types of the columns in the resultSet should be left to the evaluator of the parsed expressions...  so I will probably add a new method of a parsed expression, probably named "typeOf", that takes an environment (where each name is mapped to just some value of the correct type) and returns a type from java.sql.Types...