105, return SQL NULL from ResultSet.get*() for bad/missing values
This enables bad or missing column values in the CSV file
to be identified and handled by the application.

Change StringConverter conversion methods to return null if numeric
or date values not parseable (instead of returning 0 for numbers and
'1970-01-01' for dates).

Adapt CsvResult.get*() methods to return SQL NULL or zero for
columns that cannot be parsed.

Added new test file bad_values.csv and
unit test TestCsvDriver.testBadColumnValues