130, use US locale for converting column type names to upper/lower
Use Locale.US for converting column type names and date/time/timestamp
formats to upper/lowercase.

In Turkish, lowercase of "I" is not "i", so safe solution is to use
American Locale.US in String.toLowerCase()/toUpperCase(), as all of
these strings are always ASCII, and Locale.US is guaranteed to be
available.

See http://mattryall.net/blog/2009/02/the-infamous-turkish-locale-bug

No unit test added because test would depend on the CI server
allowing default Locale to be temporarily changed to Turkish.